requirements.txt
supabase
python-dotenv
pandas
requests
